+++ Dockerfile
@@ -10,4 +10,4 @@ COPY requirements.txt /tmp/req.txt
 
 RUN pip install -r /tmp/req.txt
 
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
