+++ app/models/department.py
@@ -1,0 +1,15 @@
+from pydantic import BaseModel
+from typing import Optional
+from datetime import datetime
+
+class DepartmentBase(BaseModel):
+    name: str
+    description: str
+
+class DepartmentPatch(BaseModel):
+    name: Optional[str] = None
+    description: Optional[str] = None
+
+class Department(DepartmentBase):
+    id: str
+    createdAt: datetime
