+++ app/models/user.py
@@ -1,0 +1,21 @@
+from pydantic import BaseModel
+from typing import Optional
+from datetime import datetime
+
+class UserBase(BaseModel):
+    username: str
+    fullName: str
+
+class UserCreate(UserBase):
+    password: str
+
+class User(UserBase):
+    id: str
+    created_at: datetime
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str = "bearer"
+
+class TokenData(BaseModel):
+    username: Optional[str] = None
