+++ app/routers/auth.py
@@ -3,13 +3,13 @@ from fastapi.security import OAuth2PasswordRequestForm
 from datetime import timedelta
 from ..utils.auth import (
     verify_password,
-    get_password_hash,
+    hash_password as get_password_hash,
     create_access_token,
     ACCESS_TOKEN_EXPIRE_MINUTES
 )
 from ..config.database import read_db, write_db
-from ..models.schemas import UserCreate, Token
-from ..utils.logger import logger
+from ..models.user import UserCreate, Token
+from ..utils.logger import logger, log_failed_login, log_successful_login
 import time
 
 router = APIRouter()
@@ -40,6 +40,7 @@ async def login(form_data: OAuth2PasswordRequestForm = Depends()):
     user = next((user for user in users if user["username"] == form_data.username), None)
 
     if not user or not verify_password(form_data.password, user["password"]):
+        log_failed_login(form_data.username, "Invalid credentials")
         raise HTTPException(
             status_code=401,
             detail="Incorrect username or password",
@@ -50,5 +51,5 @@ async def login(form_data: OAuth2PasswordRequestForm = Depends()):
     access_token = create_access_token(
         data={"sub": user["username"]}, expires_delta=access_token_expires
     )
-    logger.info(f"User logged in: {form_data.username}")
+    log_successful_login(form_data.username)
     return {"access_token": access_token, "token_type": "bearer"}
